nrp cleanup:
1) use RtpHeader.FIXED_HEADER_SIZE_BYTES instead of one defined in NRP
2) Get rid of now-unused methods
3) Move readRegionToBuf to ByteArrayBuffer
